mini_project/requirements.txt
faiss-cpu==1.8.0.post1 
flask==2.3.2
flask-restful==0.3.10 
keras==2.12.0 
numpy==1.23.5
matplotlib==3.7.2
opencv-python==4.10.0.84  
opencv-python-headless==4.9.0.80
Werkzeug==2.3.6
torch==2.0.1
torchvision==0.15.2
tensorflow==2.12.0
scikit-learn==1.2.2
pandas==2.0.3
tqdm